Game.Chess.Piece

Documentation

pointValue :: Piece -> Int

Section 3, "Programming a Computer for Playing Chess" By Claude E. Shannon.

movingVectors :: Piece -> [MovingVector Int]

Determines the directions a Piece can move in one step. Does not work for pawns, since their directional vector differs depending on color. Use "movingVectors'" if you know the color.

movingVectors' :: Piece -> Color -> [MovingVector Int]

Determines the directions a Piece can go

multiMovePiece :: Piece -> Bool

Can the piece move more than one square?

piecePoints :: Board -> Color -> Int

Given a Board and a Color, determine the current piece-point value for that color.

piecePoints' :: Board -> Color -> Piece -> Int

Given a Board, a Color, and a Piece to filter for, determine the current piece-point value for that color's pieces.

materialScore :: Board -> Int

Section 3, "Programming a Computer for Playing Chess" By Claude E. Shannon.

evaluate :: Board -> Int