Game.Chess.Position

Documentation

movePosition :: (Int, Int) -> Position -> Maybe Position

Given a position vector in the form (file, rank), add it to a given Position, and return the result so long as we remain in a valid Position.

generate :: Board -> Position -> [Board]

Move generation!

allMoves :: Color -> Board -> [Board]

Determine possible moves for a given side.

movesTree :: Color -> Board -> Int -> GameTree

Generate a GameTree by calling allMoves the appropriate number of times depending on the depth.