Game.Chess.Types

Documentation

data Board

Constructors

Board

Fields

board :: Vector Cell

sideToMove :: Color

castleAbility :: [CastleAbility]

enPassant :: Maybe String

halfmoves :: Integer

fullmoves :: Integer

data CastleAbility

Constructors

Kingside Color

Queenside Color

data Piece

Constructors

Pawn

Knight

Bishop

King

Queen

Rook

data Color

Constructors

White

Black

data Cell

Constructors

Empty

Cell

Fields

piece :: Piece

color :: Color

data File

Describes a File on the board.

data Rank

Describes a Rank on the board.

pattern PRank :: Int -> Rank

pattern PFile :: Int -> File

data Position

Position allows us to discuss a position on the board, given its appropriate File and Rank. We define an Enum instance for easily converting to and from the 0x88 Vector position of a 'Board'\'s board representation.

Constructors

Position !File !Rank

getFile :: File -> Int

getRank :: Rank -> Int

Smart constructors

mkFile :: Int -> Maybe File

A smart constructor for File that ensures that the file is between 0 and 7 inclusive.

mkRank :: Int -> Maybe Rank

A smart constructor for Rank that ensures that the rank is between 0 and 7 inclusive.

mkPosition :: Maybe File -> Maybe Rank -> Maybe Position

A helper for constructing Position, useful in conjunction with the smart constructors mkFile and mkRank.