Game.Chess.Position

Documentation

movePosition :: (Int, Int) -> Position -> Maybe Position

Given a position vector in the form (file, rank), add it to a given Position, and return the result so long as we remain in a valid Position.