Game.Chess.Types

Documentation

data Board

Board represents the current state of the game at any given moment.

Constructors

Board

Fields

board :: Vector Cell

sideToMove :: Color

castleAbility :: [CastleAbility]

enPassant :: Maybe String

halfmoves :: Integer

fullmoves :: Integer

data CastleAbility

Who can castle and on which side(s)?

Constructors

Kingside Color

Queenside Color

data MovingVector a

There are certain cases where a moving vector for a piece is more than just a 2-tuple position vector. The most obvious case is the set of moving vectors for Pawns. Pawns capture diagonally, so we need a way to denote "this vector only applies when capturing."

Constructors

NormalMove a a

PawnCaptureMove a a

data Piece

There are 6 different kinds of chess pieces.

Constructors

Pawn

Knight

Bishop

King

Queen

Rook

data Color

Each piece is White or Black.

Constructors

White

Black

data Cell

A Cell represents a square on the board. It is either Empty or a Cell containing a particular Piece which belongs to a Color.

This is isomorphic to Maybe (Piece, Color).

Constructors

Empty

Cell

Fields

piece :: Piece

color :: Color

data File

Describes a File on the board.

data Rank

Describes a Rank on the board.

pattern PRank :: Int -> Rank

pattern PFile :: Int -> File

data Position

Position allows us to discuss a position on the board, given its appropriate File and Rank. We define an Enum instance for easily converting to and from the 0x88 Vector position of a 'Board'\'s board representation.

Constructors

Position !File !Rank

getFile :: File -> Int

getRank :: Rank -> Int

data GameTree

GameTree represents a tree of game states. Typically it is constructed around the same time move generation happens.

Constructors

GameTree

Fields

root :: Board

gameTree :: [GameTree]

data MoveValidity

During move generation, we need a way to describe the validity of moves.

Eventually we will need a SameSideCheck constructor to indicate that the given move would cause the current side's king to be placed in check.

Constructors

EmptySquare
We are moving to an empty cell.

Take
We are moving to a cell that has an opposite-color piece.

Occupied
We are moving to a cell that has a same-color piece.

InvalidPawnCapture
We are a Pawn wanting to capture what we cannot.

parsePosition :: String -> Maybe Position

A helper to parse positions from a String.

Smart constructors

mkFile :: Int -> Maybe File

A smart constructor for File that ensures that the file is between 0 and 7 inclusive.

mkRank :: Int -> Maybe Rank

A smart constructor for Rank that ensures that the rank is between 0 and 7 inclusive.

mkPosition :: Maybe File -> Maybe Rank -> Maybe Position

A helper for constructing Position, useful in conjunction with the smart constructors mkFile and mkRank.

Utility functions

mkMovingVector :: (a -> a -> MovingVector a) -> (a, a) -> MovingVector a

A helper function for converting from regular 2-tuples to MovingVectors.

Should be called with something like mkMovingVector (1, 1) NormalMove.

mkTuple :: MovingVector a -> (a, a)

This goes the other way, converting a MovingVector a to (a, a).

It is only here until the Position module gets updated to account for the addition of the MovingVector type. Do not rely on it.